Implemented Product classes

+++ Team-Cool_with_Coule/src/main/java/com/sequence/product/activity/DELETEProductActivity.java
@@ -2,7 +2,9 @@ package main.java.com.sequence.product.activity;
 
 import com.amazonaws.services.lambda.runtime.Context;
 import com.amazonaws.services.lambda.runtime.RequestHandler;
+import main.java.com.exception.ProductDoesNotExistException;
 import main.java.com.obj.Product;
+import main.java.com.obj.ResponseStatus;
 import main.java.com.obj.dao.ProductDao;
 import main.java.com.sequence.product.request.DELETEProductRequest;
 import main.java.com.sequence.product.result.DELETEProductResult;
@@ -21,8 +23,14 @@ public class DELETEProductActivity implements RequestHandler<DELETEProductReques
     public DELETEProductResult handleRequest(DELETEProductRequest deleteProductRequest, Context context) {
 
         Product product = dao.getProductByName(deleteProductRequest.getName());
+        if (product == null) {
+            throw new ProductDoesNotExistException("Product does not exist with the name " + deleteProductRequest.getName());
+        }
         dao.deleteProduct(product);
+        ResponseStatus status = new ResponseStatus(200, "Success");
 
-        return null;
+        return DELETEProductResult.builder()
+                .responseStatus(status)
+                .build();
     }
 }
